finished bzero, created some test.c, updated mylib.h

+++ libft/ft_memset.c
@@ -6,18 +6,18 @@
 /*   By: odruke <marvin@42lausanne.ch>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/11 09:01:30 by odruke            #+#    #+#             */
-/*   Updated: 2022/11/11 10:39:43 by odruke           ###   ########.fr       */
+/*   Updated: 2022/11/14 11:22:14 by odruke           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "mylib.h"
 
-void	*ft_memset(void *b, int c, unsigned int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
-	unsigned char *p;
-	unsigned int i;
+	unsigned char	*p;
+	unsigned int	i;
 
 	p = (unsigned char *)b;
 	i = 0;
-
 	while (i < len)
 	{
 		p[i] = c;
